requirements.txt
gradio
numpy
tensorflow-cpu